+++ CHANGELOG.md
@@ -9,6 +9,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Display samples' name (tooltip) and affected status directly on caseS page
 - Search SVs across all cases, in given genes
 - `CLINVAR_API_URL` param can be specified in app settings to override the URL used to send ClinVar submissions to. Intended for testing.
+- Support for loading and storing OMICS data
+- Parse DROP Fraser and Outrider TSVs
+- Display omics variants - wts outliers (Fraser, Outrider)
 - Hovertip to gene panel names with associated genes in SV variant view, when variant covers more than one gene
 ### Changed
 - Updated igv.js to v3.0.1